El calculo de la velocidad ya no se pierde
Habia que forzar el float en el 256. Ahora bien, al momento de
devolverlo, lo hace como negativo D: O_o

+++ TP_FINAL/proyectobase_FINAL.X/source/main.c
@@ -64,7 +64,7 @@ int main ( void )
 {
     
     int ejes;
-    unsigned int velocidad;
+    float velocidad;
     HORARIO timeStamp;
     
     unsigned int ultSec = 0;
